# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Zuu AI Video Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List user's video generations
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: favourite
          schema:
            type: boolean
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoListResponse"

    post:
      operationId: generateVideo
      tags: [videos]
      summary: Start a video generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoGenerationInput"
      responses:
        "201":
          description: Generation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoGeneration"
        "400":
          description: Validation or content error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Insufficient credits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get a video generation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video generation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoGeneration"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteVideo
      tags: [videos]
      summary: Delete a video generation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /videos/{id}/status:
    get:
      operationId: getVideoStatus
      tags: [videos]
      summary: Poll video generation status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoStatusResponse"
  /videos/{id}/regenerate:
    post:
      operationId: regenerateVideo
      tags: [videos]
      summary: Regenerate a video from its settings
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: New generation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoGeneration"
  /videos/{id}/favourite:
    post:
      operationId: toggleFavourite
      tags: [videos]
      summary: Toggle favourite status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoGeneration"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ─────────────────────────────────────────────────────────────
...
    VideoGenerationInput:
      type: object
      required: [prompt, style, aspectRatio, duration]
      properties:
        prompt:
          type: string
          minLength: 1
        enhancedPrompt:
          type: ["string", "null"]
        negativePrompt:
          type: ["string", "null"]
        style:
          type: string
        aspectRatio:
          type: string
          enum: ["9:16", "16:9", "1:1", "4:5"]
        duration:
          type: integer
          enum: [5, 10, 15, 24, 30, 60]
        quality:
          type: string
          enum: [standard, hd, full_hd, 4k]
          default: standard
        cameraMovement:
          type: ["string", "null"]
        lighting:
          type: ["string", "null"]
        motionStrength:
          type: string
          enum: [low, medium, high]
          default: medium
        seed:
          type: ["integer", "null"]
        provider:
          type: string
          default: mock
        characterProfileId:
          type: ["integer", "null"]
        referenceImageUrl:
          type: ["string", "null"]
        scenes:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/SceneInput"
    SceneInput:
      type: object
      required: [sceneNumber, description, duration]
      properties:
        sceneNumber:
          type: integer
        description:
          type: string
        characterAction:
          type: ["string", "null"]
        cameraMovement:
          type: ["string", "null"]
        duration:
          type: integer
        environment:
          type: ["string", "null"]
        dialogue:
          type: ["string", "null"]
        soundEffect:
          type: ["string", "null"]
        musicMood:
          type: ["string", "null"]
    VideoGeneration:
      type: object
      required: [id, userId, prompt, style, aspectRatio, duration, quality, status, provider, isFavourite, creditsUsed, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        prompt:
          type: string
        enhancedPrompt:
          type: ["string", "null"]
        negativePrompt:
          type: ["string", "null"]
        style:
          type: string
        aspectRatio:
          type: string
        duration:
          type: integer
        quality:
          type: string
        cameraMovement:
          type: ["string", "null"]
        lighting:
          type: ["string", "null"]
        motionStrength:
          type: string
        seed:
          type: ["integer", "null"]
        provider:
          type: string
        providerJobId:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, queued, processing, completed, failed, cancelled]
        progressMessage:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        referenceImageUrl:
          type: ["string", "null"]
        characterProfileId:
          type: ["integer", "null"]
        isFavourite:
          type: boolean
        creditsUsed:
          type: integer
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    VideoListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/VideoGeneration"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    VideoStatusResponse:
      type: object
      required: [id, status]
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [draft, queued, processing, completed, failed, cancelled]
        progressMessage:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
...